Refactor: check for dry run in subcommands

+++ install.sh
@@ -8,26 +8,36 @@
 
 install_babashka() {
     local dry_run="$1"
+    if [[ -z "$dry_run" ]]; then
+       # dry run
+       echo -n ""
+    else
     # Defer to default "babashka manual install"
     bash <(curl https://raw.githubusercontent.com/babashka/babashka/master/install)
+    fi
 }
 
 install_bbin() {
     local dry_run="$1"
-    # TODO get version first, then download that version
-    mkdir -p ~/.babashka/bbin/bin && curl -o- -L https://raw.githubusercontent.com/babashka/bbin/v0.1.2/bbin > ~/.babashka/bbin/bin/bbin && chmod +x ~/.babashka/bbin/bin/bbin
-    # TODO sanely add to PATH -- check first if we need to.
-    # if bbin path is not on PATH
-    #   if shell is bash
-    #       ...
-    #   else if shell is fish
-    #       ...
-    #   end
-    # end
-    # From bbin install instructions:
-    #   echo 'export PATH="$PATH:$HOME/.babashka/bbin/bin"' >> ~/.$(basename $SHELL)rc && exec $SHELL
-    #
-    # We need to add $HOME/.babashka/bbin/bin to PATH!
+    if [[ -z "$dry_run" ]]; then
+       # dry run
+       echo -n ""
+    else
+        # TODO get version first, then download that version
+        mkdir -p ~/.babashka/bbin/bin && curl -o- -L https://raw.githubusercontent.com/babashka/bbin/v0.1.2/bbin > ~/.babashka/bbin/bin/bbin && chmod +x ~/.babashka/bbin/bin/bbin
+        # TODO sanely add to PATH -- check first if we need to.
+        # if bbin path is not on PATH
+        #   if shell is bash
+        #       ...
+        #   else if shell is fish
+        #       ...
+        #   end
+        # end
+        # From bbin install instructions:
+        #   echo 'export PATH="$PATH:$HOME/.babashka/bbin/bin"' >> ~/.$(basename $SHELL)rc && exec $SHELL
+        #
+        # We need to add $HOME/.babashka/bbin/bin to PATH!
+    fi
 }
 
 # Just pass args into subcommands
